[FIX] merge button style components into one
버 튼 스타일 컴포넌트를 하나로 합쳤습니다
