+++ todolist/src/components/body/column/columnBody/staticCard/staticCard.jsx
@@ -5,6 +5,7 @@ import CardAuthor from './cardAuthor.jsx';
 import CardDeleteButton from './cardDeleteButton.jsx';
 import StaticCardStyle from './staticCard.style';
 import CardTextWrapper from './cardTextWrapper.style';
+import ModifyCardCancelButton from './modifyCardCancelButton';
 
 const StaticCard = ({ cardTitle, cardContents, user }) => {
   return (
@@ -14,6 +15,7 @@ const StaticCard = ({ cardTitle, cardContents, user }) => {
         <CardContents cardContents={cardContents} />
         <CardAuthor user={user} />
       </CardTextWrapper>
+      <ModifyCardCancelButton />
       <CardDeleteButton />
     </StaticCardStyle>
   );
